move legacy EventDetails
