+++ src/components/legacy/EventDetails/EventDetails.js
