+++ src/components/legacy/EventDetails/EventHeader.js
