+++ src/components/legacy/EventDetails/package.json
